Slightly re-architect alu to support improving timing
Apparently the way you model this stuff in more complicated CPUs is with a distributed cluster architecture. So you'd have a load balancer module that you instantiate in front of a set of servers (the ALU in this case) which is in charge of taking in requests and sending out responses though queues.